Move MarkedOptionsFactory to shared/lib folder
