+++ src/app/app.module.ts
@@ -38,7 +38,7 @@ import { GithubEventService } from './core/services/githubevent.service';
 import { LoggingService } from './core/services/logging.service';
 import { IssueServiceFactory } from './core/services/factories/factory.issue.service';
 import { PermissionService } from './core/services/permission.service';
-import { markedOptionsFactory } from './core/services/factories/factory.marked.options';
+import { markedOptionsFactory } from './shared/lib/marked';
 
 @NgModule({
   declarations: [
